change -> startapp chat
