+++ srcs/django_code/transcend/transcend/settings.py
@@ -44,6 +44,7 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
 	'myapp',
     'users',
+    'chat',
     'channels',
     'rest_framework_simplejwt.token_blacklist',
     'rest_framework',
